Keep portal regression within Swift file budgets
The portal regression belongs in a focused test file instead of growing TerminalAndGhosttyTests, and the runtime fix can enforce the invariant without extra debug-only lines. This keeps the large-file budget green while preserving the test-first coverage and the portal ownership change.\n\nConstraint: The Swift file length budget is enforced before launch.\nConfidence: high\nScope-risk: narrow\nTested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv\nTested: git diff --check\nTested: plutil -lint GhosttyTabs.xcodeproj/project.pbxproj\nNot-tested: XCTest not run locally per repo policy
